// frontend/src/app/model/Surete.ts
export class Surete {

    surete_numero: Number;
    surete_numpoli: Number;
    surete_numeroavenant: Number;
    surete_numeroacte: Number;
    surete_numeroengagement: Number;

    surete_typesurete: String;
    surete_identificationtitre: String;
    surete_localisation: String;
    surete_adressegeolocalisation: String;
    surete_retenudeposit: number;
    surete_datedeposit: Date;
    surete_depositlibere: number;
    surete_dateliberation: Date;
    surete_cautionsolidaire: String;
    surete_datecomptabilisation: Date;
    surete_codeutilisateur: String;
    surete_statutliberation: number;

}